Add files via upload
A project created with C++ utilizing QT. I have redacted parts that I was not part of, and those which revealed anything other than first name from my peers. At the end you can see the log table of the work I did for this project, which mostly consisted of the frontend UI functionality and design (this design is not my proudest work). I also created the backend basic functionality for a few of the classes.